fix: move Extended classes to their own separate files
