+++ lib/Service/ShiftService.php
@@ -6,14 +6,14 @@ namespace OCA\ShiftsNext\Service;
 
 use DateTimeImmutable;
 use Exception;
-use JsonSerializable;
 use OCA\ShiftsNext\Db\Shift;
 use OCA\ShiftsNext\Db\ShiftMapper;
 use OCA\ShiftsNext\Db\ShiftType;
 use OCA\ShiftsNext\Exception\ShiftNotFoundException;
 use OCA\ShiftsNext\Exception\ShiftTypeNotFoundException;
 use OCA\ShiftsNext\Exception\UserNotFoundException;
-use OCA\ShiftsNext\Serializable\SerializableUser;
+use OCA\ShiftsNext\Extended\ShiftExtended;
+use OCA\ShiftsNext\Extended\ShiftTypeExtended;
 use OCP\IUser;
 use Throwable;
 
@@ -109,35 +109,5 @@ class ShiftService {
 		} catch (Throwable) {
 			return null;
 		}
-	}
-}
-
-/**
- * This class is used to create objects representing a shift
- * with all its foreign keys resolved to full objects
- */
-final class ShiftExtended implements JsonSerializable {
-	public int $id;
-	public string $start;
-	public string $end;
-
-	public function __construct(
-		Shift $shift,
-		public IUser $user,
-		public ShiftTypeExtended $shiftType,
-	) {
-		$this->id = $shift->getId();
-		$this->start = $shift->getStart();
-		$this->end = $shift->getEnd();
-	}
-
-	public function jsonSerialize(): array {
-		return [
-			'id' => $this->id,
-			'user' => new SerializableUser($this->user),
-			'shift_type' => $this->shiftType,
-			'start' => $this->start,
-			'end' => $this->end,
-		];
 	}
 }
